Move everything to pipelines
