+++ .chloggen/mx-psi_graph.yaml
@@ -5,7 +5,7 @@ change_type: enhancement
 component: service
 
 # A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
-note: "Makes `graph` public to allow overrriding telemetry providers"
+note: "Makes `pipelines` package public to allow overrriding telemetry providers"
 
 # One or more tracking issues or pull requests related to the change
 issues: [4970]
