+++ service/pipelines/pipelines_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package graph
+package pipelines
 
 import (
 	"context"
@@ -28,7 +28,6 @@ import (
 	"go.opentelemetry.io/collector/receiver"
 	"go.opentelemetry.io/collector/receiver/receivertest"
 	"go.opentelemetry.io/collector/service/internal/testcomponents"
-	"go.opentelemetry.io/collector/service/pipelines"
 )
 
 var _ component.Component = &testNode{}
@@ -201,12 +200,12 @@ func TestGraphStartStopComponentError(t *testing.T) {
 func TestConnectorPipelinesGraph(t *testing.T) {
 	tests := []struct {
 		name                string
-		pipelineConfigs     pipelines.Config
+		pipelineConfigs     Config
 		expectedPerExporter int // requires symmetry in pipelines
 	}{
 		{
 			name: "pipelines_simple.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -227,7 +226,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_simple_mutate.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewIDWithName("exampleprocessor", "mutate")},
@@ -248,7 +247,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_simple_multi_proc.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor"), component.NewIDWithName("exampleprocessor", "mutate")},
@@ -269,7 +268,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_simple_no_proc.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers: []component.ID{component.NewID("examplereceiver")},
 					Exporters: []component.ID{component.NewID("exampleexporter")},
@@ -287,7 +286,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_multi.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver"), component.NewIDWithName("examplereceiver", "1")},
 					Processors: []component.ID{component.NewIDWithName("exampleprocessor", "mutate"), component.NewID("exampleprocessor")},
@@ -308,7 +307,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_multi_no_proc.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers: []component.ID{component.NewID("examplereceiver"), component.NewIDWithName("examplereceiver", "1")},
 					Exporters: []component.ID{component.NewID("exampleexporter"), component.NewIDWithName("exampleexporter", "1")},
@@ -326,7 +325,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "multi_pipeline_receivers_and_exporters.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewIDWithName("exampleprocessor", "mutate")},
@@ -359,7 +358,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_simple_traces.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -375,7 +374,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_simple_metrics.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -391,7 +390,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_simple_logs.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("logs", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -407,7 +406,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_fork_merge_traces.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -433,7 +432,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_fork_merge_metrics.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -459,7 +458,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_fork_merge_logs.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("logs", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -485,7 +484,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_translate_from_traces.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -506,7 +505,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_translate_from_metrics.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("metrics"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -527,7 +526,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_translate_from_logs.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewID("logs"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -548,7 +547,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_matrix.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers:  []component.ID{component.NewID("examplereceiver")},
 					Processors: []component.ID{component.NewID("exampleprocessor")},
@@ -584,7 +583,7 @@ func TestConnectorPipelinesGraph(t *testing.T) {
 		},
 		{
 			name: "pipelines_conn_lanes.yaml",
-			pipelineConfigs: pipelines.Config{
+			pipelineConfigs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers: []component.ID{component.NewID("examplereceiver")},
 					Exporters: []component.ID{component.NewID("mockforward")},
@@ -929,7 +928,7 @@ func TestConnectorRouter(t *testing.T) {
 		),
 	}
 
-	pcfg := pipelines.Config{
+	pcfg := Config{
 		tracesInID: {
 			Receivers: []component.ID{rcvrID},
 			Exporters: []component.ID{routeTracesID},
@@ -1061,7 +1060,7 @@ func TestGraphBuildErrors(t *testing.T) {
 		processorCfgs map[component.ID]component.Config
 		exporterCfgs  map[component.ID]component.Config
 		connectorCfgs map[component.ID]component.Config
-		pipelineCfgs  pipelines.Config
+		pipelineCfgs  Config
 		expected      string
 	}{
 		{
@@ -1072,7 +1071,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): badExporterFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("logs"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1088,7 +1087,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): badExporterFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("metrics"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1104,7 +1103,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): badExporterFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("traces"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1123,7 +1122,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("logs"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("bf")},
@@ -1143,7 +1142,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("metrics"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("bf")},
@@ -1163,7 +1162,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("bf")},
@@ -1180,7 +1179,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("logs"): {
 					Receivers: []component.ID{component.NewID("bf")},
 					Exporters: []component.ID{component.NewID("nop")},
@@ -1196,7 +1195,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("metrics"): {
 					Receivers: []component.ID{component.NewID("bf")},
 					Exporters: []component.ID{component.NewID("nop")},
@@ -1212,7 +1211,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("traces"): {
 					Receivers: []component.ID{component.NewID("bf")},
 					Exporters: []component.ID{component.NewID("nop")},
@@ -1231,7 +1230,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1254,7 +1253,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1277,7 +1276,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1300,7 +1299,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1323,7 +1322,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1346,7 +1345,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1369,7 +1368,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("logs", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1392,7 +1391,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("logs", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1415,7 +1414,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("bf"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("logs", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("bf")},
@@ -1438,7 +1437,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewIDWithName("nop", "conn"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewIDWithName("nop", "conn")},
@@ -1457,7 +1456,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewIDWithName("nop", "conn"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "out"): {
 					Receivers: []component.ID{component.NewIDWithName("nop", "conn")},
 					Exporters: []component.ID{component.NewID("nop")},
@@ -1476,7 +1475,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("mockforward"): mfConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("mockforward")},
@@ -1503,7 +1502,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("mockforward"): mfConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("mockforward")},
@@ -1533,7 +1532,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewIDWithName("nop", "conn"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("traces"): {
 					Receivers:  []component.ID{component.NewIDWithName("nop", "conn")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1559,7 +1558,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewIDWithName("nop", "conn"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("metrics"): {
 					Receivers:  []component.ID{component.NewIDWithName("nop", "conn")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1585,7 +1584,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewIDWithName("nop", "conn"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("logs"): {
 					Receivers:  []component.ID{component.NewIDWithName("nop", "conn")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1613,7 +1612,7 @@ func TestGraphBuildErrors(t *testing.T) {
 				component.NewIDWithName("nop", "conn1"): nopConnectorFactory.CreateDefaultConfig(),
 				component.NewIDWithName("nop", "conn2"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1658,7 +1657,7 @@ func TestGraphBuildErrors(t *testing.T) {
 				component.NewIDWithName("nop", "conn1"): nopConnectorFactory.CreateDefaultConfig(),
 				component.NewIDWithName("nop", "conn2"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("metrics", "in"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1703,7 +1702,7 @@ func TestGraphBuildErrors(t *testing.T) {
 				component.NewIDWithName("nop", "conn1"): nopConnectorFactory.CreateDefaultConfig(),
 				component.NewIDWithName("nop", "conn2"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("logs", "in"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1749,7 +1748,7 @@ func TestGraphBuildErrors(t *testing.T) {
 				component.NewIDWithName("nop", "forkagain"): nopConnectorFactory.CreateDefaultConfig(),
 				component.NewIDWithName("nop", "rawlog"):    nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -1803,7 +1802,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("traces"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("nop"), component.NewIDWithName("nop", "1")},
@@ -1819,7 +1818,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("unknown"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("traces"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("unknown")},
@@ -1838,7 +1837,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("metrics"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop"), component.NewIDWithName("nop", "1")},
@@ -1858,7 +1857,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("metrics"): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("unknown")},
@@ -1875,7 +1874,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("logs"): {
 					Receivers: []component.ID{component.NewID("nop"), component.NewIDWithName("nop", "1")},
 					Exporters: []component.ID{component.NewID("nop")},
@@ -1891,7 +1890,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			exporterCfgs: map[component.ID]component.Config{
 				component.NewID("nop"): nopReceiverFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewID("logs"): {
 					Receivers: []component.ID{component.NewID("unknown")},
 					Exporters: []component.ID{component.NewID("nop")},
@@ -1910,7 +1909,7 @@ func TestGraphBuildErrors(t *testing.T) {
 			connectorCfgs: map[component.ID]component.Config{
 				component.NewID("unknown"): nopConnectorFactory.CreateDefaultConfig(),
 			},
-			pipelineCfgs: pipelines.Config{
+			pipelineCfgs: Config{
 				component.NewIDWithName("traces", "in"): {
 					Receivers: []component.ID{component.NewID("nop")},
 					Exporters: []component.ID{component.NewID("unknown")},
@@ -2017,7 +2016,7 @@ func TestGraphFailToStartAndShutdown(t *testing.T) {
 	dataTypes := []component.DataType{component.DataTypeTraces, component.DataTypeMetrics, component.DataTypeLogs}
 	for _, dt := range dataTypes {
 		t.Run(string(dt)+"/receiver", func(t *testing.T) {
-			pcfg := pipelines.Config{
+			pcfg := Config{
 				component.NewID(dt): {
 					Receivers:  []component.ID{component.NewID("nop"), component.NewID("err")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -2031,7 +2030,7 @@ func TestGraphFailToStartAndShutdown(t *testing.T) {
 		})
 
 		t.Run(string(dt)+"/processor", func(t *testing.T) {
-			pcfg := pipelines.Config{
+			pcfg := Config{
 				component.NewID(dt): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop"), component.NewID("err")},
@@ -2045,7 +2044,7 @@ func TestGraphFailToStartAndShutdown(t *testing.T) {
 		})
 
 		t.Run(string(dt)+"/exporter", func(t *testing.T) {
-			pcfg := pipelines.Config{
+			pcfg := Config{
 				component.NewID(dt): {
 					Receivers:  []component.ID{component.NewID("nop")},
 					Processors: []component.ID{component.NewID("nop")},
@@ -2060,7 +2059,7 @@ func TestGraphFailToStartAndShutdown(t *testing.T) {
 
 		for _, dt2 := range dataTypes {
 			t.Run(string(dt)+"/"+string(dt2)+"/connector", func(t *testing.T) {
-				pcfg := pipelines.Config{
+				pcfg := Config{
 					component.NewIDWithName(dt, "in"): {
 						Receivers:  []component.ID{component.NewID("nop")},
 						Processors: []component.ID{component.NewID("nop")},
@@ -2113,7 +2112,7 @@ func (g *Pipelines) getReceivers() map[component.DataType]map[component.ID]compo
 // However, within an individual pipeline, we expect:
 // - E instances of the connector as a receiver.
 // - R instances of the connector as an exporter.
-func expectedInstances(m pipelines.Config, pID component.ID) (int, int) {
+func expectedInstances(m Config, pID component.ID) (int, int) {
 	var r, e int
 	for _, rID := range m[pID].Receivers {
 		if rID.Type() != "exampleconnector" {
